refactor processRepos() into a shared function
